+++ src/segment-loader.js
@@ -1,14 +1,17 @@
 /**
  * @file segment-loader.js
  */
-import {getMediaInfoForTime_ as getMediaInfoForTime} from './playlist';
+import Playlist from './playlist';
 import videojs from 'video.js';
 import SourceUpdater from './source-updater';
 import Config from './config';
 import window from 'global/window';
-import removeCuesFromTrack from 'videojs-contrib-media-sources/es5/remove-cues-from-track.js';
+import removeCuesFromTrack from
+  'videojs-contrib-media-sources/es5/remove-cues-from-track.js';
 import { initSegmentId } from './bin-utils';
 import {mediaSegmentRequest, REQUEST_ERRORS} from './media-segment-request';
+import { TIME_FUDGE_FACTOR, timeUntilRebuffer as timeUntilRebuffer_ } from './ranges';
+import { minRebufferMaxBandwidthSelector } from './playlist-selectors';
 import * as loadBalancer from './load-balancing';
 
 // in ms
@@ -43,6 +46,34 @@ const detectEndOfStream = function(playlist, mediaSource, segmentIndex) {
     appendedLastSegment;
 };
 
+const finite = (num) => typeof num === 'number' && isFinite(num);
+
+export const illegalMediaSwitch = (loaderType, startingMedia, newSegmentMedia) => {
+  // Although these checks should most likely cover non 'main' types, for now it narrows
+  // the scope of our checks.
+  if (loaderType !== 'main' || !startingMedia || !newSegmentMedia) {
+    return null;
+  }
+
+  if (!newSegmentMedia.containsAudio && !newSegmentMedia.containsVideo) {
+    return 'Neither audio nor video found in segment.';
+  }
+
+  if (startingMedia.containsVideo && !newSegmentMedia.containsVideo) {
+    return 'Only audio found in segment when we expected video.' +
+      ' We can\'t switch to audio only from a stream that had video.' +
+      ' To get rid of this message, please add codec information to the manifest.';
+  }
+
+  if (!startingMedia.containsVideo && newSegmentMedia.containsVideo) {
+    return 'Video found in segment when we expected only audio.' +
+      ' We can\'t switch to a stream with video from an audio only stream.' +
+      ' To get rid of this message, please add codec information to the manifest.';
+  }
+
+  return null;
+};
+
 /**
  * An object that manages segment loading and appending.
  *
@@ -51,20 +82,18 @@ const detectEndOfStream = function(playlist, mediaSource, segmentIndex) {
  * @extends videojs.EventTarget
  */
 export default class SegmentLoader extends videojs.EventTarget {
-  constructor(options) {
+  constructor(settings, options = {}) {
     super();
     // check pre-conditions
-    if (!options) {
-      throw new TypeError('Initialization options are required');
+    if (!settings) {
+      throw new TypeError('Initialization settings are required');
     }
-    if (typeof options.currentTime !== 'function') {
+    if (typeof settings.currentTime !== 'function') {
       throw new TypeError('No currentTime getter specified');
     }
-    if (!options.mediaSource) {
+    if (!settings.mediaSource) {
       throw new TypeError('No MediaSource specified');
     }
-    let settings = videojs.mergeOptions(videojs.options.hls, options);
-
     // public properties
     this.state = 'INIT';
     this.bandwidth = settings.bandwidth;
@@ -82,7 +111,9 @@ export default class SegmentLoader extends videojs.EventTarget {
     this.mediaSource_ = settings.mediaSource;
     this.hls_ = settings.hls;
     this.loaderType_ = settings.loaderType;
+    this.startingMedia_ = void 0;
     this.segmentMetadataTrack_ = settings.segmentMetadataTrack;
+    this.goalBufferLength_ = settings.goalBufferLength;
 
     // private instance variables
     this.checkBufferTimeout_ = null;
@@ -110,10 +141,12 @@ export default class SegmentLoader extends videojs.EventTarget {
 
     this.syncController_.on('syncinfoupdate', () => this.trigger('syncinfoupdate'));
 
+    this.mediaSource_.addEventListener('sourceopen', () => this.ended_ = false);
+
     // ...for determining the fetch location
     this.fetchAtBuffer_ = false;
 
-    if (settings.debug) {
+    if (options.debug) {
       this.logger_ = videojs.log.bind(videojs, 'segment-loader', this.loaderType_, '->');
     }
   }
@@ -138,6 +171,7 @@ export default class SegmentLoader extends videojs.EventTarget {
    */
   dispose() {
     this.state = 'DISPOSED';
+    this.pause();
     this.abort_();
     if (this.sourceUpdater_) {
       this.sourceUpdater_.dispose();
@@ -159,10 +193,15 @@ export default class SegmentLoader extends videojs.EventTarget {
 
     this.abort_();
 
+    // We aborted the requests we were waiting on, so reset the loader's state to READY
+    // since we are no longer "waiting" on any requests. XHR callback is not always run
+    // when the request is aborted. This will prevent the loader from being stuck in the
+    // WAITING state indefinitely.
+    this.state = 'READY';
+
     // don't wait for buffer check timeouts to begin fetching the
     // next segment
     if (!this.paused()) {
-      this.state = 'READY';
       this.monitorBuffer_();
     }
   }
@@ -194,6 +233,12 @@ export default class SegmentLoader extends videojs.EventTarget {
 
     this.pendingSegment_ = null;
     return this.error_;
+  }
+
+  endOfStream() {
+    this.ended_ = true;
+    this.pause();
+    this.trigger('ended');
   }
 
   /**
@@ -423,8 +468,10 @@ export default class SegmentLoader extends videojs.EventTarget {
    * Delete all the buffered data and reset the SegmentLoader
    */
   resetEverything() {
+    this.ended_ = false;
     this.resetLoader();
-    this.remove(0, Infinity);
+    this.remove(0, this.duration_());
+    this.trigger('reseteverything');
   }
 
   /**
@@ -445,6 +492,7 @@ export default class SegmentLoader extends videojs.EventTarget {
   resyncLoader() {
     this.mediaIndex = null;
     this.syncPoint_ = null;
+    this.abort();
   }
 
   /**
@@ -529,7 +577,7 @@ export default class SegmentLoader extends videojs.EventTarget {
                                           segmentInfo.mediaIndex);
 
     if (isEndOfStream) {
-      this.mediaSource_.endOfStream();
+      this.endOfStream();
       return;
     }
 
@@ -583,7 +631,7 @@ export default class SegmentLoader extends videojs.EventTarget {
 
     // if there is plenty of content buffered, and the video has
     // been played before relax for awhile
-    if (bufferedTime >= Config.GOAL_BUFFER_LENGTH) {
+    if (bufferedTime >= this.goalBufferLength_()) {
       return null;
     }
 
@@ -629,19 +677,19 @@ export default class SegmentLoader extends videojs.EventTarget {
     // fetch
     if (this.fetchAtBuffer_) {
       // Find the segment containing the end of the buffer
-      let mediaSourceInfo = getMediaInfoForTime(playlist,
-                                                lastBufferedEnd,
-                                                syncPoint.segmentIndex,
-                                                syncPoint.time);
+      let mediaSourceInfo = Playlist.getMediaInfoForTime(playlist,
+                                                         lastBufferedEnd,
+                                                         syncPoint.segmentIndex,
+                                                         syncPoint.time);
 
       mediaIndex = mediaSourceInfo.mediaIndex;
       startOfSegment = mediaSourceInfo.startTime;
     } else {
       // Find the segment containing currentTime
-      let mediaSourceInfo = getMediaInfoForTime(playlist,
-                                                currentTime,
-                                                syncPoint.segmentIndex,
-                                                syncPoint.time);
+      let mediaSourceInfo = Playlist.getMediaInfoForTime(playlist,
+                                                         currentTime,
+                                                         syncPoint.segmentIndex,
+                                                         syncPoint.time);
 
       mediaIndex = mediaSourceInfo.mediaIndex;
       startOfSegment = mediaSourceInfo.startTime;
@@ -720,6 +768,120 @@ export default class SegmentLoader extends videojs.EventTarget {
   }
 
   /**
+   * Determines if the network has enough bandwidth to complete the current segment
+   * request in a timely manner. If not, the request will be aborted early and bandwidth
+   * updated to trigger a playlist switch.
+   *
+   * @param {Object} stats
+   *        Object containing stats about the request timing and size
+   * @return {Boolean} True if the request was aborted, false otherwise
+   * @private
+   */
+  abortRequestEarly_(stats) {
+    if (this.hls_.tech_.paused() ||
+        // Don't abort if the current playlist is on the lowestEnabledRendition
+        // TODO: Replace using timeout with a boolean indicating whether this playlist is
+        //       the lowestEnabledRendition.
+        !this.xhrOptions_.timeout ||
+        // Don't abort if we have no bandwidth information to estimate segment sizes
+        !(this.playlist_.attributes.BANDWIDTH)) {
+      return false;
+    }
+
+    // Wait at least 1 second since the first byte of data has been received before
+    // using the calculated bandwidth from the progress event to allow the bitrate
+    // to stabilize
+    if (Date.now() - (stats.firstBytesReceivedAt || Date.now()) < 1000) {
+      return false;
+    }
+
+    const currentTime = this.currentTime_();
+    const measuredBandwidth = stats.bandwidth;
+    const segmentDuration = this.pendingSegment_.duration;
+
+    const requestTimeRemaining =
+      Playlist.estimateSegmentRequestTime(segmentDuration,
+                                          measuredBandwidth,
+                                          this.playlist_,
+                                          stats.bytesReceived);
+
+    // Subtract 1 from the timeUntilRebuffer so we still consider an early abort
+    // if we are only left with less than 1 second when the request completes.
+    // A negative timeUntilRebuffering indicates we are already rebuffering
+    const timeUntilRebuffer = timeUntilRebuffer_(this.buffered_(),
+                                                 currentTime,
+                                                 this.hls_.tech_.playbackRate()) - 1;
+
+    // Only consider aborting early if the estimated time to finish the download
+    // is larger than the estimated time until the player runs out of forward buffer
+    if (requestTimeRemaining <= timeUntilRebuffer) {
+      return false;
+    }
+
+    const switchCandidate = minRebufferMaxBandwidthSelector({
+      master: this.hls_.playlists.master,
+      currentTime,
+      bandwidth: measuredBandwidth,
+      duration: this.duration_(),
+      segmentDuration,
+      timeUntilRebuffer,
+      currentTimeline: this.currentTimeline_,
+      syncController: this.syncController_
+    });
+
+    if (!switchCandidate) {
+      return;
+    }
+
+    const rebufferingImpact = requestTimeRemaining - timeUntilRebuffer;
+
+    const timeSavedBySwitching = rebufferingImpact - switchCandidate.rebufferingImpact;
+
+    let minimumTimeSaving = 0.5;
+
+    // If we are already rebuffering, increase the amount of variance we add to the
+    // potential round trip time of the new request so that we are not too aggressive
+    // with switching to a playlist that might save us a fraction of a second.
+    if (timeUntilRebuffer <= TIME_FUDGE_FACTOR) {
+      minimumTimeSaving = 1;
+    }
+
+    if (!switchCandidate.playlist ||
+        switchCandidate.playlist.uri === this.playlist_.uri ||
+        timeSavedBySwitching < minimumTimeSaving) {
+      return false;
+    }
+
+    // set the bandwidth to that of the desired playlist being sure to scale by
+    // BANDWIDTH_VARIANCE and add one so the playlist selector does not exclude it
+    // don't trigger a bandwidthupdate as the bandwidth is artifial
+    this.bandwidth =
+      switchCandidate.playlist.attributes.BANDWIDTH * Config.BANDWIDTH_VARIANCE + 1;
+    this.abort();
+    this.trigger('earlyabort');
+    return true;
+  }
+
+  /**
+   * XHR `progress` event handler
+   *
+   * @param {Event}
+   *        The XHR `progress` event
+   * @param {Object} simpleSegment
+   *        A simplified segment object copy
+   * @private
+   */
+  handleProgress_(event, simpleSegment) {
+    if (!this.pendingSegment_ ||
+        simpleSegment.requestId !== this.pendingSegment_.requestId ||
+        this.abortRequestEarly_(simpleSegment.stats)) {
+      return;
+    }
+
+    this.trigger('progress');
+  }
+
+  /**
    * load a specific segment from a request into the buffer
    *
    * @private
@@ -734,13 +896,7 @@ export default class SegmentLoader extends videojs.EventTarget {
       this.decrypter_,
       this.createSimplifiedSegmentObj_(segmentInfo),
       // progress callback
-      (event, segment) => {
-        if (!this.pendingSegment_ || segment.requestId !== this.pendingSegment_.requestId) {
-          return;
-        }
-        // TODO: Use progress-based bandwidth to early abort low-bandwidth situations
-        this.trigger('progress');
-      },
+      this.handleProgress_.bind(this),
       this.segmentRequestFinished_.bind(this));
   }
 
@@ -757,20 +913,20 @@ export default class SegmentLoader extends videojs.EventTarget {
     const currentTime = this.currentTime_();
     let removeToTime = 0;
 
-    // Chrome has a hard limit of 150mb of
+    // Chrome has a hard limit of 150MB of
     // buffer and a very conservative "garbage collector"
     // We manually clear out the old buffer to ensure
     // we don't trigger the QuotaExceeded error
     // on the source buffer during subsequent appends
 
     // If we have a seekable range use that as the limit for what can be removed safely
-    // otherwise remove anything older than 1 minute before the current play head
+    // otherwise remove anything older than 30 seconds before the current play head
     if (seekable.length &&
         seekable.start(0) > 0 &&
         seekable.start(0) < currentTime) {
       removeToTime = seekable.start(0);
     } else {
-      removeToTime = currentTime - 60;
+      removeToTime = currentTime - 30;
     }
 
     if (removeToTime > 0) {
@@ -848,6 +1004,7 @@ export default class SegmentLoader extends videojs.EventTarget {
     // an error occurred from the active pendingSegment_ so reset everything
     if (error) {
       this.pendingSegment_ = null;
+      this.state = 'READY';
 
       // the requests were aborted just record the aborted stat and exit
       // this is not a true error condition and nothing corrective needs
@@ -857,7 +1014,6 @@ export default class SegmentLoader extends videojs.EventTarget {
         return;
       }
 
-      this.state = 'READY';
       this.pause();
 
       // the error is really just that at least one of the requests timed-out
@@ -932,12 +1088,36 @@ export default class SegmentLoader extends videojs.EventTarget {
       return;
     }
 
-    this.state = 'APPENDING';
-
     const segmentInfo = this.pendingSegment_;
     const segment = segmentInfo.segment;
+    const timingInfo = this.syncController_.probeSegmentInfo(segmentInfo);
 
-    this.syncController_.probeSegmentInfo(segmentInfo);
+    // When we have our first timing info, determine what media types this loader is
+    // dealing with. Although we're maintaining extra state, it helps to preserve the
+    // separation of segment loader from the actual source buffers.
+    if (typeof this.startingMedia_ === 'undefined' &&
+        timingInfo &&
+        // Guard against cases where we're not getting timing info at all until we are
+        // certain that all streams will provide it.
+        (timingInfo.containsAudio || timingInfo.containsVideo)) {
+      this.startingMedia_ = {
+        containsAudio: timingInfo.containsAudio,
+        containsVideo: timingInfo.containsVideo
+      };
+    }
+
+    const illegalMediaSwitchError =
+      illegalMediaSwitch(this.loaderType_, this.startingMedia_, timingInfo);
+
+    if (illegalMediaSwitchError) {
+      this.error({
+        message: illegalMediaSwitchError
+      });
+      this.trigger('error');
+      return;
+    }
+
+    this.state = 'APPENDING';
 
     if (segmentInfo.isSyncRequest) {
       this.trigger('syncinfoupdate');
@@ -949,6 +1129,8 @@ export default class SegmentLoader extends videojs.EventTarget {
     if (segmentInfo.timestampOffset !== null &&
         segmentInfo.timestampOffset !== this.sourceUpdater_.timestampOffset()) {
       this.sourceUpdater_.timestampOffset(segmentInfo.timestampOffset);
+      // fired when a timestamp offset is set in HLS (can also identify discontinuities)
+      this.trigger('timestampoffset');
     }
 
     // if the media initialization segment is changing, append it
@@ -1040,7 +1222,7 @@ export default class SegmentLoader extends videojs.EventTarget {
                                             segmentInfo.mediaIndex + 1);
 
     if (isEndOfStream) {
-      this.mediaSource_.endOfStream();
+      this.endOfStream();
     }
 
     if (!this.paused()) {
@@ -1098,6 +1280,11 @@ export default class SegmentLoader extends videojs.EventTarget {
     const segment = segmentInfo.segment;
     const start = segment.start;
     const end = segment.end;
+
+    // Do not try adding the cue if the start and end times are invalid.
+    if (!finite(start) || !finite(end)) {
+      return;
+    }
 
     removeCuesFromTrack(start, end, this.segmentMetadataTrack_);